refactor(api): decompose OpenBaoCluster type definitions
Organize the OpenBaoCluster API declarations by domain while preserving the existing CRD and generated schema surface.
